refactor: 식물 가이드 상세 메시지 Provider를 여러 Provider로 분리
