+++ src/main/kotlin/dnd11th/blooming/api/service/guide/StaticPlantMessageFactory.kt
@@ -2,7 +2,7 @@ package dnd11th.blooming.api.service.guide
 
 import dnd11th.blooming.api.dto.guide.PlantGuideDetailViewResponse
 import dnd11th.blooming.api.dto.guide.PlantGuideSimpleViewResponse
-import dnd11th.blooming.api.service.guide.provider.StaticPlantDetialMessageProvider
+import dnd11th.blooming.api.service.guide.provider.StaticPlantDetailMessageProvider
 import dnd11th.blooming.api.service.guide.provider.StaticPlantSimpleMessageProvider
 import dnd11th.blooming.domain.entity.plant.Plant
 import org.springframework.stereotype.Component
@@ -11,7 +11,7 @@ import java.time.Month
 @Component
 class StaticPlantMessageFactory(
     private val simpleMessageProvider: StaticPlantSimpleMessageProvider,
-    private val detailMessageProvider: StaticPlantDetialMessageProvider,
+    private val detailMessageProvider: StaticPlantDetailMessageProvider,
 ) : PlantMessageFactory {
     override fun buildTags(plant: Plant): List<String> {
         val tagList = mutableListOf<String>()
