+++ src/main/kotlin/dnd11th/blooming/api/service/guide/provider/ToxicityDetailProvider.kt
@@ -1,0 +1,33 @@
+package dnd11th.blooming.api.service.guide.provider
+
+import dnd11th.blooming.api.dto.guide.DetailToxicityResponse
+import dnd11th.blooming.common.util.KoreanUtil.Companion.getEunOrNun
+import dnd11th.blooming.domain.entity.plant.Plant
+import dnd11th.blooming.domain.entity.plant.Toxicity
+import org.springframework.stereotype.Component
+
+@Component
+class ToxicityDetailProvider {
+    companion object {
+        const val TOXICITY_TITLE = "독성"
+    }
+
+    fun createDetailToxicityResponse(plant: Plant): DetailToxicityResponse {
+        return DetailToxicityResponse(
+            title = TOXICITY_TITLE,
+            description = makeDetailToxicityDescription(plant),
+        )
+    }
+
+    private fun makeDetailToxicityDescription(plant: Plant): String {
+        val eunOrNun = getEunOrNun(plant.korName)
+
+        val description =
+            when (plant.toxicity) {
+                Toxicity.NOT_EXISTS -> "독성이 없어요."
+                Toxicity.EXISTS -> "독성이 있어 반려동물과 어린아이들이 먹지 않도록 주의해야 해요."
+            }
+
+        return "${plant.korName}$eunOrNun $description"
+    }
+}
