--- src/main/kotlin/dnd11th/blooming/common/exception/DeleteException.kt
@@ -1,3 +1,0 @@
-package dnd11th.blooming.common.exception
-
-class DeleteException(errorType: ErrorType) : MyException(errorType)
